fix: seed gifts with explicit IDs to avoid BIGSERIAL null constraint error
Neon.tech's gifts table didn't have the BIGSERIAL sequence set up,
causing seed to fail. Now provides explicit IDs (1-8) and uses
ON CONFLICT (id) DO NOTHING for idempotency.

+++ scripts/migrate.js
@@ -91,7 +91,7 @@ const migrate = async () => {
     // ── Gifts ────────────────────────────────────────────────────────────────
     await client.query(`
       CREATE TABLE IF NOT EXISTS gifts (
-        id         BIGSERIAL PRIMARY KEY,
+        id         INTEGER PRIMARY KEY,
         name       VARCHAR(50) NOT NULL,
         emoji      VARCHAR(10) NOT NULL,
         price      DECIMAL(8,2) NOT NULL,
